Removed some tests from date_picker_test as they are now covered in the calendar and input date picker tests.
